==> research/01_data_ingestion.ipynb <==
import os

# os.chdir('../')
%pwd
# os.chdir('../')

from dataclasses import dataclass
from pathlib import Path

@dataclass(frozen=True)
class DataIngestionConfig:
    root_dir: Path
    kaggle_dataset: str
    local_data_file: Path
    unzip_dir: Path


from pathlib import Path
from typing import Any, Dict

from wine_quality_predictor.utils.common import read_yaml, make_directory
from wine_quality_predictor.constants import *


class ConfigurationManager:
    def __init__(
        self,
        config_filepath: Path = CONFIG_FILE_PATH,
        params_filepath: Path = PARAMS_FILE_PATH,
        schema_filepath: Path = SCHEMA_FILE_PATH
    ):
        self.config_filepath = config_filepath
        # self.params_filepath = params_filepath
        # self.schema_filepath = schema_filepath

        self.config = read_yaml(Path(self.config_filepath))
        # self.params = read_yaml(self.params_filepath)
        # self.schema = read_yaml(self.schema_filepath)

        make_directory(Path(self.config.artifacts_root))

    def get_data_ingestion_config(self) -> DataIngestionConfig:
        config = self.config.data_ingestion

        make_directory(config.root_dir)

        return DataIngestionConfig(
            root_dir=config.root_dir,
            kaggle_dataset=config.kaggle_dataset,
            local_data_file=config.local_data_file,
            unzip_dir=config.unzip_dir
        )


import os
import subprocess
import zipfile
from wine_quality_predictor.utils.common import get_file_size
import logging


class DataIngestion:
    def __init__(self, config: DataIngestionConfig):
        self.config = config

    def download_data(self):
        if not os.path.exists(self.config.local_data_file):
            logging.info("Downloading dataset from Kaggle...")
            command = [
                "kaggle", "datasets", "download",
                "-d", self.config.kaggle_dataset,
                "-p", str(self.config.root_dir),
                "--force"
            ]
            subprocess.run(command, check=True)
            logging.info(f"Downloaded dataset to {self.config.local_data_file}")
        else:
            logging.info("Dataset already downloaded.")

    def extract_zip_file(self):
        unzip_path = self.config.unzip_dir
        os.makedirs(unzip_path, exist_ok=True)

        with zipfile.ZipFile(self.config.local_data_file, 'r') as zip_ref:
            zip_ref.extractall(unzip_path)

        logging.info(f"Extracted zip file to: {unzip_path}")
        logging.info(f"Files extracted: {os.listdir(unzip_path)}")


try:
        logging.info(f">>>>>> Stage Data Ingestion started <<<<<<")
        
        data_ingestion_config = ConfigurationManager()
        config = data_ingestion_config.get_data_ingestion_config()
        data_ingestion = DataIngestion(config)
        data_ingestion.download_data()
        data_ingestion.extract_zip_file()

        logging.info(f">>>>>> Stage Data Ingestion completed <<<<<<\n")

except Exception as e:
    logging.exception(f"Error occurred in stage Data Ingestion: {e}")
    raise e
